# Conflicts:
#	src/components/assetsManager/AddPlatformDialog/AddPlatformDialog.jsx
#	src/components/assetsManager/AddTokenDialog/AddTokenDialog.jsx
#	src/components/assetsManager/AssetManagerDialog/AssetManagerDialog.jsx
#	src/components/assetsManager/AssetManagerDialog/AssetManagerForm.jsx
#	src/components/assetsManager/RevokeDialog/RevokeDialog.jsx
#	src/components/dashboard/SendTokens/SendTokensForm.jsx
#	src/components/dialogs/AddCurrencyDialog.jsx
#	src/components/dialogs/AddTokenDialog.jsx
#	src/components/dialogs/CBEAddressDialog.jsx
#	src/components/dialogs/CBETokenDialog.jsx
#	src/components/dialogs/ConfirmTxDialog/ConfirmTxDialog.js
#	src/components/dialogs/CopyDialog.jsx
#	src/components/dialogs/ModalDialog.jsx
#	src/components/dialogs/ModalDialogBase/ModalDialogBase.js
#	src/components/dialogs/PollDialog.jsx
#	src/components/dialogs/VoteDialog.jsx
#	src/components/dialogs/wallet/WalletAddEditDialog/WalletAddEditDialog.jsx
#	src/components/dialogs/wallet/WalletAddEditDialog/WalletAddEditForm.jsx
#	src/components/dialogs/wallet/WalletSelectDialog.jsx
#	src/dao/MultisigWalletDAO.js
#	src/layouts/partials/InfoPartial/InfoPartial.jsx
#	src/models/notices/WalletNoticeModel.js
#	src/redux/assetsManager/actions.js
#	src/redux/market/action.js
#	src/redux/modals/actions.js
#	src/redux/session/actions.js
#	src/redux/wallet/actions.js
#	src/redux/watcher/actions.js
#	src/services/MultisigWalletService.js

+++ src/dao/MultisigWalletDAO.js
@@ -1,12 +1,11 @@
 import BigNumber from 'bignumber.js'
-
-import MultisigWalletModel from 'models/Wallet/MultisigWalletModel'
-import MultisigWalletPendingTxCollection from 'models/Wallet/MultisigWalletPendingTxCollection'
 import MultisigWalletPendingTxModel from 'models/Wallet/MultisigWalletPendingTxModel'
 import TokenModel from 'models/TokenModel'
-
+import MultisigWalletModel from 'models/Wallet/MultisigWalletModel'
 import AbstractMultisigContractDAO from './AbstractMultisigContractDAO'
 import contractManagerDAO from './ContractsManagerDAO'
+import MultisigWalletPendingTxCollection from 'models/Wallet/MultisigWalletPendingTxCollection'
+import MultisigTransactionModel from 'models/Wallet/MultisigTransactionModel'
 
 const CODE_CONFIRMATION_NEEDED = 14014
 
@@ -22,46 +21,62 @@ export default class MultisigWalletDAO extends AbstractMultisigContractDAO {
   }
 
   watchOwnerRemoved (wallet, callback) {
-    return this._watch('OwnerRemoved', callback, { self: wallet.address() })
+    return this._watch('MultisigWalletOwnerRemoved', callback, { self: wallet.address() })
   }
 
   watchConfirmationNeeded (wallet, callback) {
-    return this._watch('ConfirmationNeeded', (result) => {
-      const { operation, value, to, symbol } = result.args
+    return this._watch('MultisigWalletConfirmationNeeded', result => {
+      const {operation, value, to, symbol} = result.args
       const symbolString = this._c.bytesToString(symbol)
+      if (!symbolString) {
+        // eslint-disable-next-line
+        console.error('symbol not found', symbolString)
+        return
+      }
       const tokenDAO = wallet.tokens().get(symbolString).dao()
       callback(new MultisigWalletPendingTxModel({
         id: operation,
         value: tokenDAO.removeDecimals(value),
         to,
         symbol: symbolString,
-        isSigned: true,
+        isConfirmed: true,
       }))
     }, { self: wallet.address() })
   }
 
   watchMultiTransact (wallet, callback) {
-    return this._watch('MultiTransact', callback, { self: wallet.address() })
+    return this._watch('MultisigWalletMultiTransact', result => {
+      const {operation, owner, self, symbol, value} = result.args
+      const symbolString = this._c.bytesToString(symbol)
+      const token = wallet.tokens().get(symbolString)
+      callback(new MultisigTransactionModel({
+        id: operation,
+        owner,
+        wallet: self,
+        symbol: symbolString,
+        value: token.dao().removeDecimals(value),
+      }))
+    }, { self: wallet.address() })
   }
 
   watchSingleTransact (wallet, callback) {
-    return this._watch('SingleTransact', callback, { self: wallet.address() })
+    return this._watch('MultisigWalletSingleTransact', callback, { self: wallet.address() })
   }
 
   watchDeposit (wallet, callback) {
-    return this._watch('Deposit', (result) => {
+    return this._watch('MultisigWalletDeposit', result => {
       callback(wallet.tokens().get('ETH').dao().removeDecimals(result.args.value))
     }, { self: wallet.address() })
   }
 
   watchRevoke (wallet, callback) {
-    return this._watch('Revoke', (result) => {
+    return this._watch('MultisigWalletRevoke', result => {
       callback(result.args.operation)
     }, { self: wallet.address() })
   }
 
   watchConfirmation (wallet, callback) {
-    return this._watch('Confirmation', (result) => {
+    return this._watch('MultisigWalletConfirmation', result => {
       // TODO @dkchv: something wrong with contract
       if (result.args.owner === '0x') {
         return
@@ -72,16 +87,21 @@ export default class MultisigWalletDAO extends AbstractMultisigContractDAO {
 
   async getPendings (tokens) {
     let pendingTxCollection = new MultisigWalletPendingTxCollection()
-    const [to, value, symbol, id] = await this._call('getPendings')
+    const [to, value, symbol, id, isConfirmed] = await this._call('getPendings')
 
     to.forEach((item, i) => {
       const symbolString = this._c.bytesToString(symbol[i])
+      if (!symbolString) {
+        // TODO @dkchv: something wrong in contract
+        return
+      }
       const tokenDAO = tokens.get(symbolString).dao()
       pendingTxCollection = pendingTxCollection.add(new MultisigWalletPendingTxModel({
         to: item,
         value: tokenDAO.removeDecimals(value[i]),
         symbol: symbolString,
         id: id[i],
+        isConfirmed: isConfirmed[i],
       }))
     })
     return pendingTxCollection
@@ -107,7 +127,7 @@ export default class MultisigWalletDAO extends AbstractMultisigContractDAO {
   }
 
   async removeWallet (wallet, account: string) {
-    const result = await this._multisigTx('removeWallet', [
+    const result = await this._tx('removeWallet', [
       account,
     ], {
       address: wallet.address(),
@@ -116,12 +136,22 @@ export default class MultisigWalletDAO extends AbstractMultisigContractDAO {
     return result.tx
   }
 
-  async addOwner (wallet, newOwner) {
+  async addOwner (wallet, ownerAddress) {
     const result = await this._tx('addOwner', [
-      newOwner,
+      ownerAddress,
     ], {
       wallet: wallet.address(),
-      newOwner,
+      ownerAddress,
+    })
+    return result.tx
+  }
+
+  async removeOwner (wallet, ownerAddress) {
+    const result = await this._tx('removeOwner', [
+      ownerAddress,
+    ], {
+      wallet: wallet.address(),
+      ownerAddress,
     })
     return result.tx
   }
